pyproject.toml
[project]
name = "audio-tools"
version = "0.1.0"
description = "Audio transcription and article generation via Aliyun DashScope paraformer-v2"
requires-python = ">=3.10"
dependencies = [
    "dashscope>=1.20.0",
    "oss2>=2.18.0",
    "python-dotenv>=1.0.0",
]
